# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mystic Weave",
    "description": "Persistent game state backend for a text-based RPG powered by a custom GPT. The GPT is the narrator; this API is the memory. d100 roll-under resolution with domain scores and competency tiers.",
    "version": "5.3.1"
  },
  "servers": [
    {
      "url": "https://mysticweave-production.up.railway.app",
      "description": "Production (Railway)"
    }
  ],
  "paths": {
...
    "/character/{session_id}/spend_ap": {
      "post": {
        "summary": "Spend Ap",
        "description": "Spend AP to raise a domain score. AP is drawn from the single fungible pool (points_available). Bracket costs apply point-by-point: 1 AP per point at scores 25-60, 2 AP at 61-70, 3 AP at 71-80.",
        "operationId": "spend_ap_character__session_id__spend_ap_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpendAPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendAPResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/companion/new": {
      "post": {
        "summary": "Create Companion",
        "operationId": "create_companion_companion_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/companion/{companion_id}/transition": {
      "post": {
        "summary": "Transition Companion",
        "operationId": "transition_companion_companion__companion_id__transition_post",
        "parameters": [
          {
            "name": "companion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Companion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionCompanionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/companion/{companion_id}": {
      "get": {
        "summary": "Get Companion",
        "operationId": "get_companion_companion__companion_id__get",
        "parameters": [
          {
            "name": "companion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Companion Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvancementState": {
        "properties": {
          "points_available": {
            "type": "integer",
            "title": "Points Available",
            "default": 0
          },
          "points_spent": {
            "type": "integer",
            "title": "Points Spent",
            "default": 0
          },
          "points_earned_total": {
            "type": "integer",
            "title": "Points Earned Total",
            "default": 0
          },
          "tag_counter": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Tag Counter",
            "default": 0
          }
        },
        "type": "object",
        "title": "AdvancementState",
        "description": "Fungible AP pool, lifetime totals, and a single tag advance counter.\n\nTag advances increment tag_counter. Every 3 advances (counter == 3)\nresets the counter to 0 and adds 1 to points_available. Awarded AP\ngrants drop directly into points_available. Spend draws from\npoints_available with bracketed cost (1/2/3 by score range).\n\nAll four fields are server-authoritative and recomputed during state\ndelta merges. Clients submitting these values via /state/{id}/delta\nor /state/{id} will have them overwritten by the server-side\nrecompute. The fields appear on inbound payloads only for round-trip\nsafety — treat them as effectively read-only from the client's\nperspective. AP is earned (counter rollover, awarded grants) and\nspent (`/character/{id}/spend_ap`); it is not authored."
      },
...
      "CompanionResponse": {
        "properties": {
          "companion_id": {
            "type": "string",
            "title": "Companion Id"
          },
          "companion": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SapientCompanion-Output"
              },
              {
                "$ref": "#/components/schemas/CreatureCompanion"
              },
              {
                "$ref": "#/components/schemas/ExceptionalCompanion-Output"
              }
            ],
            "title": "Companion",
            "discriminator": {
              "propertyName": "tier",
              "mapping": {
                "creature": "#/components/schemas/CreatureCompanion",
                "exceptional": "#/components/schemas/ExceptionalCompanion-Output",
                "sapient": "#/components/schemas/SapientCompanion-Output"
              }
            }
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "companion_id",
          "companion"
        ],
        "title": "CompanionResponse"
      },
...
      "CreateCompanionRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "handler_id": {
            "type": "string",
            "title": "Handler Id"
          },
          "tier": {
            "type": "string",
            "enum": [
              "sapient",
              "creature",
              "exceptional"
            ],
            "title": "Tier"
          },
          "companion": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SapientCompanion-Input"
              },
              {
                "$ref": "#/components/schemas/CreatureCompanion"
              },
              {
                "$ref": "#/components/schemas/ExceptionalCompanion-Input"
              }
            ],
            "title": "Companion",
            "discriminator": {
              "propertyName": "tier",
              "mapping": {
                "creature": "#/components/schemas/CreatureCompanion",
                "exceptional": "#/components/schemas/ExceptionalCompanion-Input",
                "sapient": "#/components/schemas/SapientCompanion-Input"
              }
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "handler_id",
          "tier",
          "companion"
        ],
        "title": "CreateCompanionRequest"
      },
...
      "CreatureCompanion": {
        "properties": {
          "tier": {
            "type": "string",
            "const": "creature",
            "title": "Tier",
            "default": "creature"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "subspecies": {
            "type": "string",
            "title": "Subspecies"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "size": {
            "type": "string",
            "enum": [
              "tiny",
              "small",
              "medium",
              "large",
              "huge"
            ],
            "title": "Size"
          },
          "size_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Note"
          },
          "age_category": {
            "type": "string",
            "enum": [
              "juvenile",
              "young_adult",
              "adult",
              "mature",
              "elder"
            ],
            "title": "Age Category"
          },
          "tactical_roles": {
            "items": {
              "type": "string",
              "enum": [
                "mount",
                "pack",
                "scout",
                "guard",
                "hunter",
                "companion"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Tactical Roles"
          },
          "training_level": {
            "type": "string",
            "enum": [
              "untrained",
              "basic",
              "trained",
              "expert"
            ],
            "title": "Training Level"
          },
          "bond_level": {
            "type": "string",
            "enum": [
              "wary",
              "accepting",
              "bonded",
              "devoted"
            ],
            "title": "Bond Level"
          },
          "natural_abilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Natural Abilities"
          },
          "learned_commands": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Learned Commands"
          },
          "command_notes": {
            "type": "string",
            "title": "Command Notes",
            "default": ""
          },
          "movement_modes": {
            "items": {
              "type": "string",
              "enum": [
                "walk",
                "fly",
                "swim",
                "climb",
                "burrow"
              ]
            },
            "type": "array",
            "title": "Movement Modes"
          },
          "natural_weapons": {
            "items": {
              "type": "string",
              "enum": [
                "bite",
                "claw",
                "hoof",
                "tail_slam",
                "breath",
                "sting",
                "none"
              ]
            },
            "type": "array",
            "title": "Natural Weapons"
          },
          "carrying_capacity": {
            "type": "string",
            "enum": [
              "none",
              "small",
              "medium",
              "large"
            ],
            "title": "Carrying Capacity",
            "default": "none"
          },
          "hp": {
            "$ref": "#/components/schemas/HP"
          },
          "domains": {
            "$ref": "#/components/schemas/CreatureDomains"
          },
          "temperament": {
            "type": "string",
            "title": "Temperament",
            "default": ""
          },
          "bond_links": {
            "$ref": "#/components/schemas/BondLinks"
          },
          "narrative": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreatureNarrative"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "species",
          "subspecies",
          "size",
          "age_category",
          "tactical_roles",
          "training_level",
          "bond_level",
          "hp",
          "domains",
          "bond_links"
        ],
        "title": "CreatureCompanion"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SapientCompanion-Input": {
        "properties": {
          "tier": {
            "type": "string",
            "const": "sapient",
            "title": "Tier",
            "default": "sapient"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ancestry": {
            "type": "string",
            "title": "Ancestry"
          },
          "culture": {
            "type": "string",
            "title": "Culture"
          },
          "background": {
            "type": "string",
            "title": "Background"
          },
          "focus": {
            "type": "string",
            "title": "Focus"
          },
          "hp": {
            "$ref": "#/components/schemas/HP"
          },
          "domains": {
            "$ref": "#/components/schemas/DomainScores"
          },
          "knowledge": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Knowledge"
          },
          "application": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Application"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity-Input"
          },
          "equipment": {
            "$ref": "#/components/schemas/Equipment-Input"
          },
          "reputation": {
            "items": {
              "$ref": "#/components/schemas/ReputationEntry"
            },
            "type": "array",
            "title": "Reputation"
          },
          "bond_links": {
            "$ref": "#/components/schemas/BondLinks"
          },
          "known_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Languages"
          },
          "companions": {
            "items": {
              "$ref": "#/components/schemas/CreatureCompanion"
            },
            "type": "array",
            "title": "Companions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "ancestry",
          "culture",
          "background",
          "focus",
          "hp",
          "domains",
          "identity",
          "equipment",
          "bond_links"
        ],
        "title": "SapientCompanion"
      },
      "SapientCompanion-Output": {
        "properties": {
          "tier": {
            "type": "string",
            "const": "sapient",
            "title": "Tier",
            "default": "sapient"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ancestry": {
            "type": "string",
            "title": "Ancestry"
          },
          "culture": {
            "type": "string",
            "title": "Culture"
          },
          "background": {
            "type": "string",
            "title": "Background"
          },
          "focus": {
            "type": "string",
            "title": "Focus"
          },
          "hp": {
            "$ref": "#/components/schemas/HP"
          },
          "domains": {
            "$ref": "#/components/schemas/DomainScores"
          },
          "knowledge": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Knowledge"
          },
          "application": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Application"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity-Output"
          },
          "equipment": {
            "$ref": "#/components/schemas/Equipment-Output"
          },
          "reputation": {
            "items": {
              "$ref": "#/components/schemas/ReputationEntry"
            },
            "type": "array",
            "title": "Reputation"
          },
          "bond_links": {
            "$ref": "#/components/schemas/BondLinks"
          },
          "known_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Languages"
          },
          "companions": {
            "items": {
              "$ref": "#/components/schemas/CreatureCompanion"
            },
            "type": "array",
            "title": "Companions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "ancestry",
          "culture",
          "background",
          "focus",
          "hp",
          "domains",
          "identity",
          "equipment",
          "bond_links"
        ],
        "title": "SapientCompanion"
      },
...
      "SpendAPRequest": {
        "properties": {
          "target_domain": {
            "type": "string",
            "title": "Target Domain",
            "description": "Domain to advance. Must be one of the seven domains."
          },
          "points": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Points",
            "description": "Number of domain points to purchase. Must be positive."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_domain",
          "points"
        ],
        "title": "SpendAPRequest",
        "description": "Body for POST /character/{session_id}/spend_ap"
      },
      "SpendAPResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "target_domain": {
            "type": "string",
            "title": "Target Domain"
          },
          "points_purchased": {
            "type": "integer",
            "title": "Points Purchased"
          },
          "ap_cost_total": {
            "type": "integer",
            "title": "Ap Cost Total"
          },
          "new_domain_score": {
            "type": "integer",
            "title": "New Domain Score"
          },
          "advancement": {
            "$ref": "#/components/schemas/AdvancementState"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "target_domain",
          "points_purchased",
          "ap_cost_total",
          "new_domain_score",
          "advancement"
        ],
        "title": "SpendAPResponse"
      },
...
      "TransitionCompanionRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "new_companion": {
            "$ref": "#/components/schemas/ExceptionalCompanion-Input"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "new_companion",
          "trigger"
        ],
        "title": "TransitionCompanionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}